+++ NEWS.md
@@ -13,6 +13,7 @@
  
 ### Bug Fixes
  * Fixed a bug that was failing when wrapping and section dividers were used at the same time.
+ * Fixed a bug in `as_result_df` causing misalignment of column names.
 
 ### Miscellaneous
  * Applied `styler` and resolved package lint. Changed default indentation from 4 spaces to 2.